improve phpspec, prevent removing item from automatic right after is added to list without criteria

+++ src/SWP/Bundle/CoreBundle/EventListener/AddArticleToListListener.php
@@ -97,7 +97,7 @@ class AddArticleToListListener
                 continue;
             }
 
-            if (null !== $item) {
+            if (null !== $item && count($filters) > 0 && !$this->articleCriteriaMatcher->match($article, new Criteria($filters))) {
                 $this->contentListItemRepository->remove($item);
                 $contentList->setUpdatedAt(new \DateTime());
             }
